feat: 반복되는 API 호출 로직 useFetch 훅으로 분리
