+++ src/pages/Home/Home.tsx
@@ -5,52 +5,23 @@ import { useNavigate } from 'react-router-dom';
 import QuestionCardSection from './components/sections/QuestionCardSection';
 import BeforeAnswerSection from './components/sections/BeforeAnswerSection';
 import AnsweringSection from './components/sections/AnsweringSection';
-import useUser from './hooks/useUser';
+import useFetch from '../../shared/hooks/useFetch';
 
 export type AnswerType = 'voice' | 'text' | null;
 export type AnswerStateType = 'before-answer' | 'answering' | 'answered';
-
-// const UserSchema = z.object({
-//   user: z.object({
-//     user_id: z.number(),
-//     name: z.string(),
-//   }),
-// });
-
-// type User = z.infer<typeof UserSchema>['user'];
 
 const HomePage = () => {
   const [answerType, setAnswerType] = useState<AnswerType>(null);
   const [answerState, setAnswerState] = useState<AnswerStateType>('before-answer');
   const navigate = useNavigate();
   // const { id } = useParams();
-  const { user } = useUser();
+  const { data: user } = useFetch('/api/user', { params: { userId: 1 } });
   // TODO: 추후 콘솔 삭제
   console.log(user);
 
   const handleAnswerTypeChange = (e: ChangeEvent<HTMLInputElement>) => {
     setAnswerType(e.target.value as AnswerType);
   };
-
-  // const handleAnswerDone = async (answerText: string) => {
-  // if (answerText.trim() === '') {
-  //   alert('답변을 입력해주세요.');
-  //   return;
-  // }
-  // try {
-  //   // TODO: 로딩 상태 UI 처리
-  //   const userId = 1;
-  //   const questionId = 12345;
-  //   const postResponse = await postAnswer(userId, questionId, answerText);
-  //   const { feedbackId } = postResponse;
-  //   const feedbackData = await getFeedback(feedbackId);
-  //   navigate(ROUTE_PATH.FEEDBACK, { state: { result: feedbackData } });
-  // } catch (error) {
-  //   console.error('API 처리 실패:', error);
-  //   alert('오류가 발생했습니다. 다시 시도해주세요.');
-  //   // TODO: 로딩 상태 UI 처리
-  // }
-  // };
 
   const handleAnswerDone = () => {
     setAnswerState('answered');
